ci: run in github actions
