+++ packages/benchmark/benchmark/results/reduce.chart.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" />
     <script src="https://cdn.jsdelivr.net/npm/chart.js@3.5.1/dist/chart.min.js"></script>
-    <title>Parsing</title>
+    <title>Parser construction + one parse</title>
     <style>
       body {
         margin: 0;
@@ -28,7 +28,7 @@
   </head>
   <body>
     <div class="container">
-      <canvas id="chart1730075385339" width="16" height="9"></canvas>
+      <canvas id="chart1730217718984" width="16" height="9"></canvas>
     </div>
     <script>
       const format = (num) => {
@@ -51,18 +51,18 @@
           chunked.map((chunk) => chunk.join('')).join(' ') + fractionStr
         )
       }
-      const ctx1730075385339 = document
-        .getElementById('chart1730075385339')
+      const ctx1730217718984 = document
+        .getElementById('chart1730217718984')
         .getContext('2d')
-      const chart1730075385339 = new Chart(ctx1730075385339, {
+      const chart1730217718984 = new Chart(ctx1730217718984, {
         type: 'bar',
         data: {
           labels: ["JIT compilation","non-JIT compilation"],
           datasets: [
             {
-              data: [36784529,6311143],
-              backgroundColor: ["hsl(120, 85%, 55%)","hsl(20.592, 85%, 55%)"],
-              borderColor: ["hsl(120, 85%, 55%)","hsl(20.592, 85%, 55%)"],
+              data: [128293,2436883],
+              backgroundColor: ["hsl(6.312000000000006, 85%, 55%)","hsl(120, 85%, 55%)"],
+              borderColor: ["hsl(6.312000000000006, 85%, 55%)","hsl(120, 85%, 55%)"],
               borderWidth: 2,
             },
           ],
@@ -72,7 +72,7 @@
           plugins: {
             title: {
               display: true,
-              text: 'Parsing',
+              text: 'Parser construction + one parse',
               font: { size: 20 },
               padding: 20,
             },
