+++ packages/benchmark/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "benchmarks",
+  "name": "benchmark",
   "version": "1.0.0",
   "main": "index.js",
   "type": "module",
