+++ packages/benchmark/src/index.js
@@ -1,0 +1,5 @@
+import { printCpuInfo } from './printCpuInfo.js'
+import { compiledSuite } from './compiled.js'
+
+printCpuInfo()
+compiledSuite()
