+++ packages/benchmark/src/printCpuInfo.js
