--- packages/benchmarks/benchmark/results/reduce.json
@@ -1,27 +1,0 @@
-{
-  "name": "Parsing",
-  "date": "2024-10-28T00:29:45.339Z",
-  "version": "1.0.0",
-  "results": [
-    {
-      "name": "JIT compilation",
-      "ops": 36784529,
-      "margin": 1.78,
-      "percentSlower": 0
-    },
-    {
-      "name": "non-JIT compilation",
-      "ops": 6311143,
-      "margin": 0.43,
-      "percentSlower": 82.84
-    }
-  ],
-  "fastest": {
-    "name": "JIT compilation",
-    "index": 0
-  },
-  "slowest": {
-    "name": "non-JIT compilation",
-    "index": 1
-  }
-}
